implement authorization

+++ src/services/notes/controller/note-controller.js
@@ -1,10 +1,12 @@
 import NoteRepositories from '../repository/notes-repositories.js';
-import { InvariantError, NotFoundError } from '../../../exceptions/index.js';
+import { InvariantError, NotFoundError, AuthorizationError } from '../../../exceptions/index.js';
 import response from '../../../utils/response.js';
 
 export const createNote = async (req, res, next) => {
   const { title, body, tags } = req.validated;
-  const note = await NoteRepositories.createNote({ title, body, tags });
+  const { id: owner } = req.user;
+
+  const note = await NoteRepositories.createNote({ title, body, tags, owner });
 
   if (!note) {
     return next(new InvariantError('Catatan gagal ditambahkan'));
@@ -14,14 +16,21 @@ export const createNote = async (req, res, next) => {
 };
 
 export const getNotes = async (req, res) => {
-  const notes = await NoteRepositories.getNotes();
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
   return response(res, 200, 'Catatan sukses ditampilkan', { notes });
 };
 
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const note = await NoteRepositories.getNoteById(id);
- 
   if (!note) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
@@ -32,7 +41,13 @@ export const getNoteById = async (req, res, next) => {
 export const editNoteById = async (req, res, next) => {
   const { id } = req.params;
   const { title, body, tags } = req.validated;
- 
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const note = await NoteRepositories.editNote({ id, title, body, tags });
   if (!note) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
@@ -43,6 +58,13 @@ export const editNoteById = async (req, res, next) => {
 
 export const deleteNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const deletedNote = await NoteRepositories.deleteNote(id);
  
   if (!deletedNote) {
